Updated script to work with Python3
The script previously only worked with Python2 but the updated script now supports Python3. The main changes are the renaming of urllib2 to urllib.request, encoding changes because the program uses nordic characters, examples of this are lines 55 & 95 this line encodes the URL correctly so that the urlopen() method works on the next line. Other changes include adding parentheses around the print statements.